Move all decoders to src/model/decoder.ts
