+++ src/model/serlo.ts
@@ -24,10 +24,9 @@ import * as t from 'io-ts'
 import fetch, { Response } from 'node-fetch'
 import * as R from 'ramda'
 
-import { CommentDecoder, UuidPayloadDecoder } from './decoder'
+import { CommentDecoder, InstanceDecoder, UuidPayloadDecoder } from './decoder'
 import { Environment } from '~/internals/environment'
 import { createHelper, createMutation, createQuery } from '~/internals/model'
-import { InstanceDecoder } from '~/schema/instance/decoder'
 import { isInstance } from '~/schema/instance/utils'
 import {
   AbstractNotificationEventPayload,
