+++ src/schema/uuid/abstract-exercise/utils.ts
@@ -21,7 +21,7 @@
  */
 import { AbstractExercisePayload } from './types'
 import { requestsOnlyFields, Resolver } from '~/internals/graphql'
-import { SolutionDecoder } from '~/schema/uuid/solution/decoder'
+import { SolutionDecoder } from '~/model'
 import { SolutionPayload } from '~/schema/uuid/solution/types'
 
 export interface ExerciseResolvers<E extends AbstractExercisePayload> {
