+++ src/schema/uuid/exercise/resolvers.ts
@@ -20,16 +20,13 @@
  * @link      https://github.com/serlo-org/api.serlo.org for the canonical source repository
  */
 import { ExercisePayload, ExerciseRevisionPayload } from './types'
+import { ExerciseDecoder, ExerciseRevisionDecoder } from '~/model'
 import { createExerciseResolvers } from '~/schema/uuid/abstract-exercise/utils'
 import {
   createRepositoryResolvers,
   createRevisionResolvers,
 } from '~/schema/uuid/abstract-repository/utils'
 import { createTaxonomyTermChildResolvers } from '~/schema/uuid/abstract-taxonomy-term-child/utils'
-import {
-  ExerciseDecoder,
-  ExerciseRevisionDecoder,
-} from '~/schema/uuid/exercise/decoder'
 
 export const resolvers = {
   Exercise: {
