+++ src/schema/uuid/user/resolvers.ts
@@ -30,12 +30,12 @@ import {
   ErrorEvent,
 } from '~/internals/error-event'
 import { Context } from '~/internals/graphql'
+import { UserDecoder } from '~/model'
 import { CellValues, MajorDimension } from '~/model/google-spreadsheet-api'
 import { ConnectionPayload } from '~/schema/connection/types'
 import { resolveConnection } from '~/schema/connection/utils'
 import { createThreadResolvers } from '~/schema/thread/utils'
 import { createUuidResolvers } from '~/schema/uuid/abstract-uuid/utils'
-import { UserPayloadDecoder } from '~/schema/uuid/user/decoder'
 
 export const resolvers: UserResolvers = {
   Query: {
@@ -96,7 +96,7 @@ async function resolveUserConnectionFromIds({
       try {
         return await context.dataSources.model.serlo.getUuidWithCustomDecoder({
           id,
-          decoder: UserPayloadDecoder,
+          decoder: UserDecoder,
         })
       } catch (e) {
         return null
